Adjustments to respond and add interaction.edit
